ajustes nas estruturas de tarefas e atualizações

+++ abito-arcano/src/components/ListaTarefas.jsx
@@ -5,30 +5,26 @@ import { addItem, updateItem, toggleFinalizada, deleteItem } from './todoUtils';
 import { substituirTarefasGerais } from '../auth/firebaseTarefas';
 import { getDias, inserirDias } from '../auth/firebaseDiasHoras';
 
-function ListaTarefas({ user, tarefas, setTarefas, setPontuacoes, setDias, dias, tarefasPorDia, setTarefasPorDia }) {
+function ListaTarefas({ user, tarefas, setTarefas, setPontuacoes, setDias, dias, tarefasPorDia, setTarefasPorDia, areas }) {
   const [novaTarefa, setNovaTarefa] = useState('');
   const [itemEditando, setItemEditando] = useState(null);
 
   const resetarListaDeTarefasGerais = async (userId) => {
     try {
-      // 1. Substituir a lista de tarefas gerais com um array vazio.
       const tarefasVazias = [];
       await substituirTarefasGerais(userId, tarefasVazias);
       setTarefas(tarefasVazias)
 
-      // 2. Obter a lista de dias do Firebase.
       const dias = await getDias(userId);
 
       console.log("dias")
       console.log(dias)
 
-      // 3. Esvaziar o array de tarefas de todos os dias.
       const diasAtualizados = dias.map(dia => ({
         ...dia,
         tarefas: []
       }));
 
-      // 4. Inserir a lista de dias atualizada no Firebase.
       await inserirDias({ userId, dias: diasAtualizados });
 
       console.log("Lista de tarefas gerais e tarefas de todos os dias resetadas com sucesso.");
@@ -53,14 +49,14 @@ function ListaTarefas({ user, tarefas, setTarefas, setPontuacoes, setDias, dias,
         onChange={(e) => setNovaTarefa(e.target.value)}
         placeholder="Digite o nome da tarefa"
       />
-      <button onClick={() => addItem(novaTarefa, 'tarefa', setTarefas, tarefas, user.uid, setDias, dias, tarefasPorDia, setTarefasPorDia)}>Adicionar Tarefa</button>
+      <button onClick={() => addItem(novaTarefa, 'tarefa', setTarefas, tarefas, user.uid, setDias, dias, tarefasPorDia, setTarefasPorDia, areas)}>Adicionar Tarefa</button>
       <ul>
         {tarefas.filter(tarefa => !tarefa.finalizada).map((tarefa) => (
           <li key={tarefa.id}>
             <Tarefa
               tarefa={tarefa}
               onEdit={() => setItemEditando(tarefa)}
-              onDelete={() => deleteItem(tarefa.id, 'tarefa', setTarefas, tarefas)}
+              onDelete={() => deleteItem(tarefa.id, 'tarefa', setTarefas, tarefas, user.uid, setDias, dias, tarefasPorDia, setTarefasPorDia)}
               onToggle={() => toggleFinalizada(tarefa.id, 'tarefa', tarefas, setTarefas, setPontuacoes, user.uid)}
             />
           </li>
@@ -70,9 +66,10 @@ function ListaTarefas({ user, tarefas, setTarefas, setPontuacoes, setDias, dias,
         <EditorItem
           item={itemEditando}
           setItemEditando={setItemEditando}
-          onSave={(nome, numero, area, subarea) => updateItem(itemEditando.id, nome, numero, area, subarea, 'tarefa', setTarefas, tarefas, user.uid, setDias, dias, tarefasPorDia, setTarefasPorDia)}
+          onSave={(nome, numero, area, subarea, areaId, subareaId) => updateItem(itemEditando.id, nome, numero, area, subarea, areaId, subareaId, 'tarefa', setTarefas, tarefas, user.uid, setDias, dias, tarefasPorDia, setTarefasPorDia, areas)}
         />
       )}
+
       <h2>Finalizadas</h2>
       <ul>
         {tarefas.filter(tarefa => tarefa.finalizada).map((tarefa) => (
